fix: battery of fixes requested in PR (common activity row, and minor other changes)
